Fix layering violation from #220073
That PR introduced circular dependencies LLVMTransformUtils <-> LLVMPasses. Move Trigger*CrashPasses into LLVMPasses.
