--- llvm/include/llvm/CodeGen/TriggerCrashMachineFunction.h
@@ -1,25 +1,0 @@
-//===----------------------------------------------------------------------===//
-//
-// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM Exceptions
-//
-//===----------------------------------------------------------------------===//
-
-#ifndef LLVM_CODEGEN_TRIGGERCRASHMACHINEFUNCTION_H
-#define LLVM_CODEGEN_TRIGGERCRASHMACHINEFUNCTION_H
-
-#include "llvm/CodeGen/MachinePassManager.h"
-
-namespace llvm {
-
-class TriggerCrashMachineFunctionPass
-    : public OptionalPassInfoMixin<TriggerCrashMachineFunctionPass> {
-public:
-  LLVM_ABI PreservedAnalyses run(MachineFunction &MF,
-                                 MachineFunctionAnalysisManager &MFAM);
-};
-
-} // namespace llvm
-
-#endif // LLVM_CODEGEN_TRIGGERCRASHMACHINEFUNCTION_H
