+++ llvm/lib/CodeGen/TargetPassConfig.cpp
@@ -50,7 +50,6 @@
 #include "llvm/Transforms/ObjCARC.h"
 #include "llvm/Transforms/Scalar.h"
 #include "llvm/Transforms/Utils.h"
-#include "llvm/Transforms/Utils/TriggerCrashPass.h"
 #include <cassert>
 #include <optional>
 #include <string>
@@ -101,6 +100,21 @@ static cl::opt<bool> DisableCGP("disable-cgp", cl::Hidden,
 static cl::opt<bool>
     TriggerCrash("codegen-pipeline-trigger-crash", cl::init(false), cl::Hidden,
                  cl::desc("Trigger crash in codegen pipeline"));
+
+namespace {
+class TriggerCrashFunctionLegacyPass : public FunctionPass {
+public:
+  static char ID;
+  TriggerCrashFunctionLegacyPass() : FunctionPass(ID) {}
+  bool runOnFunction(Function &F) override {
+    abort();
+    return false;
+  }
+  StringRef getPassName() const override { return "TriggerCrashFunctionPass"; }
+};
+} // namespace
+
+char TriggerCrashFunctionLegacyPass::ID = 0;
 
 static cl::opt<bool> DisableCopyProp("disable-copyprop", cl::Hidden,
     cl::desc("Disable Copy Propagation pass"));
@@ -1090,7 +1104,7 @@ bool TargetPassConfig::addISelPasses() {
   addIRPasses();
 
   if (TriggerCrash)
-    addPass(createTriggerCrashFunctionPass());
+    addPass(new TriggerCrashFunctionLegacyPass());
 
   addCodeGenPrepare();
   addPassesToHandleExceptions();
