+++ llvm/lib/Passes/CMakeLists.txt
@@ -4,6 +4,7 @@ add_llvm_component_library(LLVMPasses
   PassBuilderBindings.cpp
   PassBuilderPipelines.cpp
   StandardInstrumentations.cpp
+  TriggerCrashPasses.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${LLVM_MAIN_INCLUDE_DIR}/llvm
