+++ llvm/lib/Passes/TriggerCrashPasses.cpp
@@ -6,18 +6,12 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "llvm/Transforms/Utils/TriggerCrashPass.h"
+#include "llvm/Passes/TriggerCrashPasses.h"
 
 using namespace llvm;
 
 PreservedAnalyses TriggerCrashModulePass::run(Module &,
                                               ModuleAnalysisManager &) {
-  abort();
-  return PreservedAnalyses::all();
-}
-
-PreservedAnalyses TriggerCrashFunctionPass::run(Function &,
-                                                FunctionAnalysisManager &) {
   abort();
   return PreservedAnalyses::all();
 }
@@ -30,6 +24,12 @@ PreservedAnalyses TriggerCrashCGSCCPass::run(LazyCallGraph::SCC &,
   return PreservedAnalyses::all();
 }
 
+PreservedAnalyses TriggerCrashFunctionPass::run(Function &,
+                                                FunctionAnalysisManager &) {
+  abort();
+  return PreservedAnalyses::all();
+}
+
 PreservedAnalyses TriggerCrashLoopPass::run(Loop &, LoopAnalysisManager &,
                                             LoopStandardAnalysisResults &,
                                             LPMUpdater &) {
@@ -37,21 +37,9 @@ PreservedAnalyses TriggerCrashLoopPass::run(Loop &, LoopAnalysisManager &,
   return PreservedAnalyses::all();
 }
 
-namespace {
-class TriggerCrashFunctionLegacyPass : public FunctionPass {
-public:
-  static char ID;
-  TriggerCrashFunctionLegacyPass() : FunctionPass(ID) {}
-  bool runOnFunction(Function &F) override {
-    abort();
-    return false;
-  }
-  StringRef getPassName() const override { return "TriggerCrashFunctionPass"; }
-};
-} // namespace
-
-char TriggerCrashFunctionLegacyPass::ID = 0;
-
-FunctionPass *llvm::createTriggerCrashFunctionPass() {
-  return new TriggerCrashFunctionLegacyPass();
+PreservedAnalyses
+TriggerCrashMachineFunctionPass::run(MachineFunction &,
+                                     MachineFunctionAnalysisManager &) {
+  abort();
+  return PreservedAnalyses::all();
 }
