+++ llvm/utils/gn/secondary/llvm/lib/CodeGen/BUILD.gn
@@ -249,7 +249,6 @@ static_library("CodeGen") {
     "TargetRegisterInfo.cpp",
     "TargetSchedule.cpp",
     "TargetSubtargetInfo.cpp",
-    "TriggerCrashMachineFunction.cpp",
     "TwoAddressInstructionPass.cpp",
     "TypePromotion.cpp",
     "UnreachableBlockElim.cpp",
