+++ llvm/utils/gn/secondary/llvm/lib/Transforms/Utils/BUILD.gn
@@ -101,7 +101,6 @@ static_library("Utils") {
     "StripGCRelocates.cpp",
     "StripNonLineTableDebugInfo.cpp",
     "SymbolRewriter.cpp",
-    "TriggerCrashPass.cpp",
     "UnifyLoopExits.cpp",
     "Utils.cpp",
     "VNCoercion.cpp",
